Refactor queries to split site and user
