+++ client/dashboard/app/queries/site-purchases.ts
@@ -1,7 +1,7 @@
 import { queryOptions } from '@tanstack/react-query';
 import { fetchSitePurchases } from '../../data/site-purchases';
 
-export const siteHasCancelablePurchasesQuery = ( siteId: number, userId: number ) =>
+export const siteHasCancelablePurchasesQuery = ( siteId: string, userId: string ) =>
 	queryOptions( {
 		queryKey: [ 'site', siteId, 'purchases', 'has-cancelable' ],
 		queryFn: () => fetchSitePurchases( siteId ),
@@ -10,7 +10,7 @@ export const siteHasCancelablePurchasesQuery = ( siteId: number, userId: number 
 				.filter( ( purchase ) => {
 					// Exclude inactive purchases and legacy premium theme purchases.
 					if (
-						! purchase.active ||
+						purchase.subscription_status !== 'active' ||
 						purchase.expiry_status === 'expired' ||
 						purchase.product_slug === 'premium_theme'
 					) {
@@ -19,15 +19,21 @@ export const siteHasCancelablePurchasesQuery = ( siteId: number, userId: number 
 
 					return purchase.is_cancelable;
 				} )
-				.filter( ( purchase ) => Number( purchase.user_id ) === userId );
+				.filter( ( purchase ) => String( purchase.user_id ) === userId );
 
 			return cancelables.length > 0;
 		},
 	} );
 
-export const sitePurchaseQuery = ( siteId: number, purchaseId: string ) =>
+export const sitePurchaseQuery = ( siteId: string, purchaseId: number ) =>
 	queryOptions( {
 		queryKey: [ 'site', siteId, 'purchases', purchaseId ],
 		queryFn: () => fetchSitePurchases( siteId ),
-		select: ( purchases ) => purchases.find( ( p ) => p.ID === purchaseId ),
+		select: ( purchases ) => purchases.find( ( p ) => String( p.ID ) === String( purchaseId ) ),
+	} );
+
+export const sitePurchasesQuery = ( siteId: string ) =>
+	queryOptions( {
+		queryKey: [ 'me', 'purchases', siteId ],
+		queryFn: () => fetchSitePurchases( siteId ),
 	} );
