+++ client/dashboard/data/site-purchases.ts
@@ -1,18 +1,9 @@
 import wpcom from 'calypso/lib/wp';
+import type { Purchase } from './purchase';
 
-export interface Purchase {
-	ID: string;
-	active: boolean;
-	expiry_status: 'manual-renew' | 'active' | 'expired';
-	expiry_message: string;
-	is_cancelable: boolean;
-	partner_name: string;
-	product_slug: string;
-	user_id: string;
-}
-
-export async function fetchSitePurchases( siteId: number ): Promise< Purchase[] > {
-	return wpcom.req.get( {
+export async function fetchSitePurchases( siteId: number | string ): Promise< Purchase[] > {
+	return await wpcom.req.get( {
 		path: `/sites/${ siteId }/purchases`,
+		apiVersion: '1.1',
 	} );
 }
